+++ client/src/App.js
@@ -7,17 +7,17 @@ import {
 // import RegistrationForm from "./components/RegistrationForm";
 // import LoginForm from "./components/LoginForm";
 import LogoutButton from "./components/LogoutButton";
-import { BrowserRouter, Switch } from "react-router-dom";
-
+import { BrowserRouter, Switch, Route } from "react-router-dom";
 import "./App.module.css";
 import Main from "./pages/Main";
 import ModalContainer from "./components/ModalContainer";
+// import PrivateRoute from "./components/PrivateRoute";
 
 function App() {
 	useAuthTokenStore();
 
 	const isAuthenticated = useIsAuthenticated();
-	const authenticatedUser = useAuthenticatedUser();
+	// const authenticatedUser = useAuthenticatedUser();
 	const [loginModal, setLoginModal] = useState(false);
 	const [user, setUser] = useState(null);
