+++ client/src/components/MemberModalBox/MemberModalBox.module.css
@@ -91,26 +91,6 @@
 	border: 2px solid;
 }
 
-.memberBox .snsBox {
-	text-align: center;
-}
-
-.memberBox .snsBox > button {
-	width: 52px;
-	height: 52px;
-	border-radius: 50%;
-	margin-bottom: 0px;
-}
-
-.memberBox .snsBox > p {
-	margin-bottom: 10px;
-	color: #666;
-}
-
-.memberBox .snsBox > button + button {
-	margin-left: 10px;
-}
-
 .btnClose {
 	width: 24px;
 	height: 24px;
